web_interface/backend/requirements.txt
# ==================== 核心框架 ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ==================== 数据库相关 ====================
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# ==================== 缓存和队列 ====================
redis==5.0.1
celery==5.3.4
flower==2.0.1

# ==================== 安全认证 ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# ==================== 数据验证和序列化 ====================
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# ==================== HTTP客户端 ====================
httpx==0.25.2
aiofiles==23.2.1

# ==================== 配置管理 ====================
python-decouple==3.8

# ==================== 日志系统 ====================
loguru==0.7.2
structlog==23.2.0

# ==================== 数据处理 ====================
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# ==================== 工具库 ====================
click==8.1.7
typer==0.9.0
rich==13.7.0
python-dateutil==2.8.2

# ==================== 监控和调试 ====================
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# ==================== 开发工具 ====================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# ==================== 测试数据生成 ====================
factory-boy==3.3.0
faker==20.1.0

# ==================== 部署相关 ====================
gunicorn==21.2.0